+++ src/delineator/__init__.py
@@ -1,4 +1,4 @@
-from .core import delineate as delineate, downloader as downloader
+from .core import delineate as delineate, downloader as downloader, clear_cache as clear_cache
 from .settings import DelineatorConfig as DelineatorConfig
 from .util import write_outputs as write_outputs
 from .serve import serve as serve